\name{generateBaseChars-methods}
\docType{methods}
\alias{generateBaseChars-methods}
\alias{generateBaseChars,PopModelSpecifier-method}
\alias{generateBaseChars}

\title{Method "generateBaseChars"  }

\description{
    generates baseline characteristics for a patient.
}

\section{Methods}{
\describe{
    \item{\code{signature(popModelSpec = "PopModelSpecifier")}}{
         generates baseline characteristics for a patient. }
}}

\keyword{methods}
