\name{sim1CT-methods}
\docType{methods}
\alias{sim1CT-methods}
\alias{sim1CT}

\alias{sim1CT,DesignSpecifier,OptionalPopModelSpecifier,OutcomeModelSpecifier-method}
\title{Method "sim1CT"  }
\description{
    simulates a single replication of a clinical trial.
}
\section{Methods}{
\describe{

    \item{\code{signature(designSpec = "DesignSpecifier", 
    popModelSpec = "OptionalPopModelSpecifier", outcomeModelSpec = "OutcomeModelSpecifier")}}{
       simulates a single replication of a clinical trial.  }
}}

\keyword{methods}

